Created distribution-deal folder and moved its content from app to lib
