--- apps/catalog/marketplace/marketplace/src/app/distribution-deal/create/create.component.html
@@ -1,161 +1,0 @@
-<div class="main" *ngIf="(movie$ | async) as movie">
-  <header>
-    <div fxLayout="column" fxLayoutAlign="center center">
-      <h1>{{movie.main.title.original}}</h1>
-      <div *ngFor="let director of movie.main.directors; let isLast = last">
-        <span>a film by {{ director.firstName }} {{ director.lastName }} <span *ngIf="!isLast">&</span>
-        </span>
-      </div>
-    </div>
-  </header>
-
-  <section fxLayout="row">
-    <aside fxLayout="row" class="page-content" [formGroup]="form">
-      <mat-card>
-        <mat-accordion multi="true">
-
-          <!-- Dates -->
-
-          <mat-expansion-panel expanded="true">
-            <mat-expansion-panel-header>
-              <mat-panel-title test-id="distribution-date-panel" class="bf-strong">
-                Dates
-              </mat-panel-title>
-            </mat-expansion-panel-header>
-            <datepicker-range [selectedRange]="choosenDateRange" [disabledDates]="occupiedDateRanges"
-              (wantedRange)="setWantedRange($event)">
-            </datepicker-range>
-          </mat-expansion-panel>
-
-          <!-- Territories -->
-
-          <mat-expansion-panel>
-            <mat-expansion-panel-header>
-              <mat-panel-title test-id="distribution-territory-panel" class="bf-strong">
-                Territories
-              </mat-panel-title>
-            </mat-expansion-panel-header>
-            <mat-form-field>
-              <mat-chip-list #chipList>
-                <mat-chip *ngFor="let territory of form.territories.value; let i = index" selectable removeable
-                  (removed)="removeTerritory(i)">
-                  {{territory | translateSlug: 'TERRITORIES'}}
-                  <mat-icon matChipRemove>cancel</mat-icon>
-                </mat-chip>
-                <input test-id="distribution-territory-input" placeholder="Territories available"
-                  [matAutocomplete]="auto" [matChipInputFor]="chipList" [formControl]="territoryControl"
-                  #territoryInput>
-              </mat-chip-list>
-              <mat-autocomplete #auto="matAutocomplete" (optionSelected)="selectedTerritory($event)">
-                <mat-option *ngFor="let territory of territoriesFilter | async" [value]="territory">
-                  {{territory | translateSlug: 'TERRITORIES'}}
-                </mat-option>
-              </mat-autocomplete>
-            </mat-form-field>
-          </mat-expansion-panel>
-
-          <!-- Media -->
-
-          <mat-expansion-panel>
-            <mat-expansion-panel-header>
-              <mat-panel-title test-id="distribution-media-panel" class="bf-strong">Mandate Media</mat-panel-title>
-            </mat-expansion-panel-header>
-            <div fxLayout="column" fxLayoutAlign="start space-between" class="filter-checkboxes-media">
-              <mat-checkbox mat-list-item *ngFor="let media of movieMedia" (change)="checkMedia(media)"
-                class="capitalized">{{media | translateSlug: 'MEDIAS'}}
-              </mat-checkbox>
-            </div>
-          </mat-expansion-panel>
-
-          <!-- Languages -->
-
-          <mat-expansion-panel>
-            <mat-expansion-panel-header>
-              <mat-panel-title test-id="languages-panel" class="bf-strong">
-                Languages
-              </mat-panel-title>
-            </mat-expansion-panel-header>
-            <form formGroupName="languages" fxLayout="column">
-              <div fxLayout="row" fxLayoutAlign="start start">
-                <mat-form-field appearance="outline">
-                  <mat-label>New language</mat-label>
-                  <input test-id="languages-panel-input" matInput type="text" [matAutocomplete]="language"
-                    [formControl]="languageControl" placeholder="Add new language" [errorStateMatcher]="matcher"
-                    #languageInput />
-                  <mat-error *ngIf="languageControl.hasError('languageNotSupported')">
-                    Please enter a valid language or choose from the dropdown menu!
-                  </mat-error>
-                  <button matSuffix mat-icon-button (click)="addLanguage(languageControl.value)"
-                    [disabled]="!languageControl.valid">
-                    <mat-icon>search</mat-icon>
-                  </button>
-                  <mat-autocomplete #language="matAutocomplete" (optionSelected)="addLanguage(languageControl.value)">
-                    <mat-option *ngFor="let language of languagesFilter | async" [value]="language">
-                      {{language | translateSlug: 'LANGUAGES'}}
-                    </mat-option>
-                  </mat-autocomplete>
-                </mat-form-field>
-                <button test-id="languages" mat-button [disabled]="!languageControl.valid" color="primary"
-                  (click)="addLanguage(languageControl.value)">Add
-                  language
-                </button>
-              </div>
-              <ng-container *ngIf="form.languages.value">
-                <mat-list [formGroup]="form.languages" role="list">
-                  <ng-container *ngFor="let language of languages$ | async">
-                    <mat-list-item [formGroupName]="language">
-                      <span mat-line>
-                        <h3 class="mat-h3">{{ language | translateSlug: 'LANGUAGES' }}</h3>
-                        <button mat-icon-button (click)="removeLanguage(language)">
-                          <mat-icon class="cancel-btn">cancel</mat-icon>
-                        </button>
-                      </span>
-                      <mat-checkbox test-id="languages-checkbox-original" mat-line formControlName="original">Original
-                      </mat-checkbox>
-                      <mat-checkbox test-id="languages-checkbox-dubbed" mat-line formControlName="dubbed">Dubbed
-                      </mat-checkbox>
-                      <mat-checkbox test-id="languages-checkbox-subtitle" mat-line formControlName="subtitle">
-                        Subtitles
-                      </mat-checkbox>
-                    </mat-list-item>
-                  </ng-container>
-                </mat-list>
-              </ng-container>
-            </form>
-          </mat-expansion-panel>
-
-          <!-- Exclusive -->
-
-          <mat-expansion-panel>
-            <mat-expansion-panel-header>
-              <mat-panel-title class="bf-strong">Exclusive</mat-panel-title>
-            </mat-expansion-panel-header>
-            <div fxLayout="column" fxLayoutAlign="start space-between">
-              <mat-checkbox mat-list-item formControlName="exclusive"
-                (click)="form.get('exclusive').value = !form.get('exclusive').value">Buy exclusive distribution right
-              </mat-checkbox>
-            </div>
-          </mat-expansion-panel>
-        </mat-accordion>
-        <button test-id="distribution-search" mat-button color="primary" (click)="startResearch()"
-          [disabled]="!form.valid">Search</button>
-      </mat-card>
-    </aside>
-
-    <!-- Results -->
-
-    <article fxLayout="column" fxLayoutAlign="start center" fxLayoutGap="20px" [ngSwitch]="step">
-      <div *ngSwitchCase="steps.START">
-        <img src='assets/images/availabilities.png' />
-      </div>
-      <div *ngSwitchCase="steps.POSSIBLE">
-        <month-calendar [selectedRange]="choosenDateRange" [disabledDates]="occupiedDateRanges"
-          (rangeSelected)="changeDateFocus($event)"></month-calendar>
-        <button test-id="add-distribution-button" mat-flat-button color="accent" (click)="addDistributionDeal()">Create distribution deal</button>
-      </div>
-      <div *ngSwitchCase="steps.ERROR">
-        Error
-      </div>
-    </article>
-  </section>
-</div>
