--- apps/catalog/marketplace/marketplace/src/app/distribution-deal/create/create.component.scss
@@ -1,80 +1,0 @@
-@import 'mixins';
-
-:host {
-  display: flex;
-  flex-direction: row;
-}
-
-.page-content {
-  margin-left: 200px;
-  margin-right: 200px;
-  margin-top: 35px;
-}
-
-section {
-  width: 65%;
-  flex: 1;
-  padding: 10px;
-
-  .cancel-btn {
-    height: 18px;
-    width: 18px;
-    line-height: 18px;
-    font-size: 18px;
-  }
-
-  h3.mat-h3 {
-    display: inline;
-    font-weight: inherit;
-    font-size: 16px;
-  }
-}
-
-month-calendar {
-  width: 100%;
-}
-
-.main {
-  width: 100%;
-}
-
-header {
-  padding: 100px 0;
-  // TODO: replace image when the real one is available.
-  background-image: url('/assets/images/catalog-marketplace-banner.png');
-  background-position: center center;
-  background-attachment: fixed;
-  background-size: cover;
-  text-align: center;
-}
-
-.filter-checkboxes-media {
-  height: 260px;
-  flex-wrap: wrap;
-}
-
-.filter-checkboxes-media {
-  height: 260px;
-  flex-wrap: wrap;
-}
-
-@include mobile {
-  .page-content {
-    margin-left: 0;
-    margin-right: 0;
-  }
-}
-
-@include tablet {
-  .page-content {
-    margin-left: 50px;
-    margin-right: 50px;
-  }
-}
-
-@include medium {
-  .page-content {
-    margin-left: 50px;
-    margin-right: 50px;
-  }
-}
