--- apps/catalog/marketplace/marketplace/src/app/distribution-deal/create/create.module.ts
@@ -1,54 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { RouterModule } from '@angular/router';
-import { ReactiveFormsModule } from '@angular/forms';
-import { FlexLayoutModule } from '@angular/flex-layout';
-import { MonthCalendarModule } from '@blockframes/ui/month-calendar/month-calendar.module';
-import { DatepickerModule } from '@blockframes/ui/datepicker/datepicker.module';
-import { DistributionDealCreateComponent } from './create.component';
-import { TranslateSlugModule } from '@blockframes/utils/pipes/translate-slug.module';
-
-// Material
-import { MatListModule } from '@angular/material/list';
-import { MatFormFieldModule } from '@angular/material/form-field';
-import { MatCardModule } from '@angular/material/card';
-import { MatExpansionModule } from '@angular/material/expansion';
-import { MatAutocompleteModule } from '@angular/material/autocomplete';
-import { MatSortModule } from '@angular/material/sort';
-import { MatButtonModule } from '@angular/material/button';
-import { MatChipsModule } from '@angular/material/chips';
-import { MatInputModule } from '@angular/material/input';
-import { MatIconModule } from '@angular/material/icon';
-import { MatCheckboxModule } from '@angular/material/checkbox';
-import { MatTableModule } from '@angular/material/table';
-import { MatSnackBarModule } from '@angular/material/snack-bar';
-
-@NgModule({
-  declarations: [DistributionDealCreateComponent],
-  imports: [
-    CommonModule,
-    ReactiveFormsModule,
-    DatepickerModule,
-    MonthCalendarModule,
-    FlexLayoutModule,
-    TranslateSlugModule,
-
-    // Material
-    MatInputModule,
-    MatButtonModule,
-    MatAutocompleteModule,
-    MatExpansionModule,
-    MatCardModule,
-    MatSortModule,
-    MatChipsModule,
-    MatTableModule,
-    MatFormFieldModule,
-    MatListModule,
-    MatCheckboxModule,
-    MatIconModule,
-    MatSnackBarModule,
-
-    RouterModule.forChild([{ path: '', component: DistributionDealCreateComponent }])
-  ]
-})
-export class DistributionDealCreateModule {}
