--- apps/catalog/marketplace/marketplace/src/app/distribution-deal/selection-table/form-selection.component.html
@@ -1,77 +1,0 @@
-<table mat-table [dataSource]="dataSource" matSort class="mat-elevation-z8">
-
-        <!-- CheckBox -->
-          <ng-container matColumnDef="checkBox">
-            <th mat-header-cell *matHeaderCellDef>
-              <mat-checkbox (change)="masterToggle()"
-              [checked]="selection.hasValue() && isAllSelected"
-              [indeterminate]="selection.hasValue() && !isAllSelected"
-              aria-label="select all movies">
-              </mat-checkbox>
-            </th>
-            <td mat-cell *matCellDef="let element">
-              <mat-checkbox
-              (click)="$event.stopPropagation()"
-              (change)="toggle(element)"
-              [checked]="selection.isSelected(element)"
-              aria-label="select one movie">
-              </mat-checkbox>
-            </td>
-          </ng-container>
-
-
-          <!-- Movie name -->
-          <ng-container matColumnDef="movieName">
-              <th mat-header-cell *matHeaderCellDef mat-sort-header>Movie name</th>
-              <td mat-cell *matCellDef="let element" class="bf-strong">{{element.movieName | titlecase}}</td>
-          </ng-container>
-
-          <!-- End Rights -->
-          <ng-container matColumnDef="endRights">
-              <th mat-header-cell *matHeaderCellDef mat-sort-header>End Rights</th>
-              <td mat-cell *matCellDef="let element">{{element.endRights | titlecase}}</td>
-          </ng-container>
-
-          <!-- Territory -->
-          <ng-container matColumnDef="territory">
-              <th mat-header-cell *matHeaderCellDef mat-sort-header>Territory</th>
-              <td mat-cell *matCellDef="let element">{{element.territory | titlecase}}</td>
-          </ng-container>
-
-          <!-- Rights -->
-          <ng-container matColumnDef="rights">
-              <th mat-header-cell *matHeaderCellDef mat-sort-header>Rights</th>
-              <td mat-cell *matCellDef="let element">{{element.rights | titlecase}}</td>
-          </ng-container>
-
-          <!-- Language -->
-          <ng-container matColumnDef="languages">
-              <th mat-header-cell *matHeaderCellDef mat-sort-header>Languages</th>
-              <td mat-cell *matCellDef="let element">{{element.languages | titlecase}}</td>
-          </ng-container>
-
-          <!-- Dubbed -->
-          <ng-container matColumnDef="dubbed">
-              <th mat-header-cell *matHeaderCellDef mat-sort-header>Dubbing</th>
-              <td mat-cell *matCellDef="let element">{{element.dubbed | titlecase}}</td>
-          </ng-container>
-
-          <!-- Subtitle -->
-          <ng-container matColumnDef="subtitle">
-              <th mat-header-cell *matHeaderCellDef mat-sort-header>Subtitle</th>
-              <td mat-cell *matCellDef="let element">{{element.subtitle | titlecase}}</td>
-          </ng-container>
-
-          <!-- Trash Icon -->
-          <ng-container matColumnDef="delete">
-              <th mat-header-cell *matHeaderCellDef></th>
-              <td mat-cell *matCellDef="let element">
-                  <button mat-button (click)="remove(element)">
-                      <mat-icon>delete_outline</mat-icon>
-                  </button>
-              </td>
-          </ng-container>
-
-          <tr mat-header-row *matHeaderRowDef="columnsToDisplay"></tr>
-          <tr mat-row *matRowDef="let row; columns: columnsToDisplay;"></tr>
-      </table>
