--- apps/catalog/marketplace/marketplace/src/app/distribution-deal/selection-table/form-selection.component.scss
@@ -1,11 +1,0 @@
-table {
-    width: 100%;
-  }
-  
-  p {
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-    width: 150px;
-  }
