+++ libs/movie/src/lib/+state/movie.model.ts
@@ -5,7 +5,6 @@ import {
   MovieMain,
   MoviePromotionalDescription,
   MoviePromotionalElements,
-  DistributionDealDocumentWithDates as DistributionDeal,
   MovieSalesAgentDealDocumentWithDates as MovieSalesAgentDeal,
   MovieSalesCast,
   MovieSalesInfoDocumentWithDates as MovieSalesInfo,
@@ -14,10 +13,7 @@ import {
   Prize,
   PromotionalElement,
   Title,
-  LicenseStatus,
-  HoldbackWithDates as Holdback,
   WorkType,
-  FormatProfile,
   MovieLanguageSpecification,
   MovieLanguageTypes,
   StoreConfig,
@@ -26,9 +22,8 @@ import {
   MovieOfficialIds,
 } from './movie.firestore';
 import { createImgRef } from '@blockframes/utils/image-uploader';
-import { createTerms } from '@blockframes/utils/common-interfaces/terms';
-import { createRange } from '@blockframes/utils/common-interfaces/range';
 import { LanguagesSlug } from '../static-model';
+import { createRange } from '@blockframes/utils/common-interfaces/range';
 
 // Export for other files
 export {
@@ -48,11 +43,9 @@ export {
   MovieStory,
   MovieVersionInfo,
   Prize,
-  DistributionDealDocumentWithDates as DistributionDeal,
   MovieSalesInfoDocumentWithDates as MovieSalesInfo,
   MovieSalesAgentDealDocumentWithDates as MovieSalesAgentDeal,
-  MovieDocumentWithDates as Movie,
-  HoldbackWithDates as Holdback
+  MovieDocumentWithDates as Movie
 } from './movie.firestore';
 
 /** A factory function that creates Movie */
@@ -212,38 +205,11 @@ export function createMovieSalesAgentDeal(
   };
 }
 
-export function createDistributionDeal(params: Partial<DistributionDeal> = {}): DistributionDeal {
-  return {
-    id: '',
-    licenseStatus: LicenseStatus.unknown,
-    licenseType: [],
-    terms: createTerms(params.terms),
-    territory: [],
-    territoryExcluded: [],
-    assetLanguage: {},
-    exclusive: false,
-    titleInternalAlias: '',
-    formatProfile: FormatProfile.unknown,
-    download: false,
-    holdbacks: [],
-    catchUp: createTerms(params.catchUp),
-    ...params
-  };
-}
-
 export function createMovieBudget(params: Partial<MovieBudget> = {}): MovieBudget {
   return {
     totalBudget: '',
     ...params,
     estimatedBudget: createRange<number>(params.estimatedBudget)
-  };
-}
-
-export function createHoldback(params: Partial<Holdback> = {}): Holdback {
-  return {
-    terms: createTerms(params.terms),
-    media: '',
-    ...params
   };
 }
 
@@ -286,4 +252,15 @@ export function createOfficialIds(params: Partial<MovieOfficialIds> = {}): Movie
     isan: '',
     ...params
   };
+}
+
+export function createMovieLanguage(
+  movieLanguage: Partial<MovieLanguageSpecification> = {}
+): MovieLanguageSpecification {
+  return {
+    original: false,
+    dubbed: false,
+    subtitle: false,
+    ...movieLanguage
+  } as MovieLanguageSpecification;
 }
