+++ libs/movie/src/lib/+state/movie.service.ts
@@ -1,14 +1,11 @@
 import { Injectable } from '@angular/core';
-import { AngularFirestoreCollection } from '@angular/fire/firestore/collection/collection';
 import { AngularFirestoreDocument } from '@angular/fire/firestore/document/document';
 import { OrganizationQuery, OrganizationService, PermissionsService } from '@blockframes/organization';
 import { CollectionConfig, CollectionService, WriteOptions } from 'akita-ng-fire';
-import { firestore } from 'firebase';
-import objectHash from 'object-hash';
 import { switchMap } from 'rxjs/operators';
-import { createMovie, Movie, DistributionDeal } from './movie.model';
+import { createMovie, Movie } from './movie.model';
 import { MovieState, MovieStore } from './movie.store';
-import { Contract, createContractTitleDetail } from '@blockframes/marketplace/app/distribution-deal/+state/cart.model';
+import { Contract } from '@blockframes/marketplace/app/distribution-deal/+state/cart.model';
 import { AuthQuery } from '@blockframes/auth';
 import { createImgRef } from '@blockframes/utils/image-uploader';
 
@@ -30,8 +27,6 @@ export function cleanModel<T>(data: T): T {
 export class MovieService extends CollectionService<MovieState> {
   constructor(
     private organizationQuery: OrganizationQuery,
-    private organizationService: OrganizationService,
-    private permissionsService: PermissionsService,
     private authQuery: AuthQuery,
     store: MovieStore
   ) {
@@ -94,87 +89,10 @@ export class MovieService extends CollectionService<MovieState> {
     return this.update(id, cleanModel(movie));
   }
 
-  private movieDoc(movieId: string): AngularFirestoreDocument<Movie> {
-    return this.db.doc(`movies/${movieId}`);
-  }
-
-  /////////////////////////////
-  // CRUD DISTRIBUTION DEALS //
-  /////////////////////////////
-
-  /**
-   *
-   * @param movieId
-   */
-  private distributionDealsCollection(movieId: string): AngularFirestoreCollection<DistributionDeal> {
-    return this.movieDoc(movieId).collection('distributiondeals');
-  }
-
-  /**
-   *
-   * @param movieId
-   * @param distributionDeal
-   */
-  public async addDistributionDeal(movieId: string, distributionDeal: DistributionDeal, contract: any): Promise<string> { // @TODO (#1440) replace contract: any with contract: Contract
-    // Create an id from DistributionDeal content.
-    // A same DistributionDeal document will always have the same hash to prevent multiple insertion of same deal
-    // @TODO #1389 Use native akita-ng-fire functions : https://netbasal.gitbook.io/akita/angular/firebase-integration/collection-service
-    const dealId = objectHash(distributionDeal);
-    distributionDeal.id = dealId;
-
-    // Populate distribution deal contract
-    contract.titles[movieId] = createContractTitleDetail();
-    contract.titles[movieId].titleId = movieId;
-    contract.titles[movieId].distributionDealIds.push(dealId);
-    // @todo #1397 change this price calculus
-    contract.titles[movieId].price = contract.price;
-
-    const contractId = await this.addContract(contract);
-
-    // Link distributiondeal with contract
-    distributionDeal.contractId = contractId;
-
-    // @TODO #1389 Use native akita-ng-fire functions : https://netbasal.gitbook.io/akita/angular/firebase-integration/collection-service
-    await this.distributionDealsCollection(movieId).doc(dealId).set(distributionDeal);
-    return dealId;
-  }
-
-  /**
-   * Checks if a distribution deal is already existing for a given movie and returns it.
-   * @param movieId
-   * @param distributionDeal
-   */
-  public async existingDistributionDeal(movieId: string, distributionDeal: DistributionDeal): Promise<DistributionDeal> {
-    const dealId = objectHash(distributionDeal);
-    const distributionDealSnapshot = await this.distributionDealsCollection(movieId).doc(dealId).get().toPromise();
-    return distributionDealSnapshot.exists ? distributionDealSnapshot.data() as DistributionDeal : undefined;
-  }
-
-  /**
-   * @param movieId
-   */
-  public async getDistributionDeals(movieId: string): Promise<DistributionDeal[]> {
-    const deals = await this.distributionDealsCollection(movieId).get().toPromise();
-    return deals.docs.map(doc => this.formatDistributionDeal(doc.data()));
-  }
-
-  public formatDistributionDeal(deal: any): DistributionDeal {
-    // Dates from firebase are Timestamps, we convert it to Dates.
-    if (deal.terms.start instanceof firestore.Timestamp) {
-      deal.terms.start = deal.terms.start.toDate();
-    }
-
-    if (deal.terms.end instanceof firestore.Timestamp) {
-      deal.terms.end = deal.terms.end.toDate();
-    }
-    return deal as DistributionDeal;
-  }
-
-
   //////////////////
   /// CONTRACT STUFF
   //////////////////
-
+  // TODO: move this into specific contract lib => ISSUE#1440
   /**
    * @param contract
    */
