+++ libs/movie/src/lib/distribution-deals/+state/distribution-deal.query.spec.ts
@@ -1,0 +1,3 @@
+test('smoke', () => {
+  expect(true).toBeTruthy();
+});
