+++ libs/movie/src/lib/distribution-deals/+state/index.ts
@@ -1,0 +1,4 @@
+export * from './distribution-deal.query';
+export * from './distribution-deal.service';
+export * from './distribution-deal.store';
+export * from './distribution-deal.model';
